remixvars: add support / fix #171
allow the same remix variable to be added to the interpolation stack as long as they don't interfere with the ongoing interpolation - eg. if the new one has a delay that is larger then the left-over interpolation time of the current one. If the new variable interferes, the current one will be updated